MVP ready

+++ src/app/serverlist/serverlist.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Http } from '@angular/http';
-import { Router } from '@angular/router';
+import { Router, ActivatedRoute } from '@angular/router';
 
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/from';
@@ -21,37 +20,41 @@ export class ServerlistComponent implements OnInit {
 
   constructor(
     private router: Router,
-    private http: Http,
+    private route: ActivatedRoute,
     private serverlistService: ServerlistService,
-    private snapService: SnapRestV2Service
+    private snapService: SnapRestV2Service,
   ) { }
 
   ngOnInit() {
     this.serverlistService.getServerList()
-      .subscribe(servers => {
-        this.servers = servers;
-        Observable.from(this.servers)
-          .map(item => {
-            const server: SnapServer = <SnapServer>item;
-            server.key = item['$key'];
-            this.snapService.getTaskList(server)
-              .timeout(250)
-              .subscribe(
-                (res) => {
-                  server.available = true;
-                },
-                (err) => {
-                  server.available = false;
-                }
-              );
-            return server;
-          })
-          .subscribe();
-      });
+      .subscribe(
+        (servers) => {
+          this.servers = servers;
+          Observable.from(this.servers)
+            .map(item => {
+              const server: SnapServer = <SnapServer>item;
+              server.key = item['$key'];
+              this.snapService.getTaskList(server)
+                .timeout(250)
+                .subscribe(
+                  (res) => {
+                    server.available = true;
+                  },
+                  (err) => {
+                    server.available = false;
+                  }
+                );
+              return server;
+            })
+            .subscribe();
+        }
+      );
   }
 
   private deleteServer(server: SnapServer) {
-    this.serverlistService.removeServer(server);
+    if (confirm('Remove this server, really?')) {
+      this.serverlistService.removeServer(server);
+    }
   }
 
 }
